requirements.txt
flask==2.0.1
Werkzeug==2.0.1
numpy==1.21.0
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.3
Pillow==8.3.1
scipy==1.7.0
mediapipe==0.8.9.1
open3d==0.13.0
trimesh==3.9.35
flask-cors==3.0.10
requests==2.26.0 